daka-copilot-refactor-all-code/requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-login>=0.6.3
flask-caching>=2.1.0
flask-migrate>=4.0.5
python-telegram-bot>=20.0
apscheduler>=3.10.4
pytz>=2024.1
werkzeug>=3.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
gevent>=24.2.0
psycopg2-binary>=2.9.9
flask-jwt-extended>=4.6.0